api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Modular RAG API with JWT authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pinecone>=5.0.0",
    "google-generativeai>=0.3.0",
    "numpy>=1.25.0",
    "google-genai>=1.49.0",
    "pypdf>=6.1.3",
    "python-docx>=1.2.0",
    "beautifulsoup4>=4.14.2",
    "markdown>=3.10",
    "passlib[bcrypt]>=1.7.4",
    "reportlab>=4.0.0",
    "bs4>=0.0.2",
    "sarvamai>=0.1.22",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "fastembed>=0.2.0",
    "motor>=3.3.0",
]
